Split up all 'x' type trait includes except for constructor and assignment
