+++ src/utl/public/utl/type_traits/utl_copy_all_extents.h
@@ -5,20 +5,6 @@
 #include "utl/preprocessor/utl_namespace.h"
 
 UTL_NAMESPACE_BEGIN
-
-template <typename From, typename To>
-struct copy_extent {
-    using type = To;
-};
-template <typename From, typename To>
-struct copy_extent<From[], To> {
-    using type = To[];
-};
-template <typename From, typename To, size_t N>
-struct copy_extent<From[N], To> {
-    using type = To[N];
-};
-
 template <typename From, typename To>
 struct copy_all_extents {
     using type = To;
@@ -28,8 +14,6 @@ struct copy_all_extents<From[], To> : copy_all_extents<From, To[]> {};
 template <typename From, typename To, size_t N>
 struct copy_all_extents<From[N], To> : copy_all_extents<From, To[N]> {};
 
-template <typename F, typename T>
-using copy_extent_t = typename copy_extent<F, T>::type;
 template <typename F, typename T>
 using copy_all_extents_t = typename copy_all_extents<F, T>::type;
