+++ src/utl/public/utl/type_traits/utl_is_const.h
@@ -11,26 +11,21 @@
 UTL_NAMESPACE_BEGIN
 
 using std::is_const;
-using std::is_volatile;
 
 #  ifdef UTL_CXX17
 
 using std::is_const_v;
-using std::is_volatile_v;
 
 #  elif defined(UTL_CXX14) // ifdef UTL_CXX17
 
 template <typename T>
 UTL_INLINE_CXX17 constexpr bool is_const_v = is_const<T>::value;
-template <typename T>
-UTL_INLINE_CXX17 constexpr bool is_volatile_v = is_volatile<T>::value;
 
 #  endif // ifdef UTL_CXX17
 
 UTL_NAMESPACE_END
 
 #  define UTL_TRAIT_SUPPORTED_is_const 1
-#  define UTL_TRAIT_SUPPORTED_is_volatile 1
 
 #else // ifdef UTL_USE_STD_TYPE_TRAITS
 
@@ -39,17 +34,10 @@ UTL_NAMESPACE_END
 #  ifndef UTL_DISABLE_BUILTIN_is_const
 #    define UTL_DISABLE_BUILTIN_is_const 0
 #  endif // ifndef UTL_DISABLE_BUILTIN_is_const
-#  ifndef UTL_DISABLE_BUILTIN_is_volatile
-#    define UTL_DISABLE_BUILTIN_is_volatile 0
-#  endif // ifndef UTL_DISABLE_BUILTIN_is_volatile
 
 #  if UTL_SHOULD_USE_BUILTIN(is_const)
 #    define UTL_BUILTIN_is_const(...) __is_const(__VA_ARGS__)
 #  endif // UTL_SHOULD_USE_BUILTIN(is_const)
-
-#  if UTL_SHOULD_USE_BUILTIN(is_volatile)
-#    define UTL_BUILTIN_is_volatile(...) __is_volatile(__VA_ARGS__)
-#  endif // UTL_SHOULD_USE_BUILTIN(is_volatile)
 
 #  ifdef UTL_BUILTIN_is_const
 
@@ -85,41 +73,6 @@ UTL_NAMESPACE_END
 
 #  endif // ifdef UTL_BUILTIN_is_const
 
-#  ifdef UTL_BUILTIN_is_volatile
-
-UTL_NAMESPACE_BEGIN
-
-template <typename T>
-struct is_volatile : bool_constant<UTL_BUILTIN_is_volatile(T)> {};
-
-#    ifdef UTL_CXX14
-template <typename T>
-UTL_INLINE_CXX17 constexpr bool is_volatile_v = UTL_BUILTIN_is_volatile(T);
-#    endif // UTL_CXX14
-
-UTL_NAMESPACE_END
-
-#  else // ifdef UTL_BUILTIN_is_volatile
-
-UTL_NAMESPACE_BEGIN
-
-template <typename T>
-struct is_volatile : false_type {};
-template <typename T>
-struct is_volatile<T volatile> : true_type {};
-template <typename T>
-struct is_volatile<T const volatile> : true_type {};
-
-#    ifdef UTL_CXX14
-template <typename T>
-UTL_INLINE_CXX17 constexpr bool is_volatile_v = is_volatile<T>::value;
-#    endif // UTL_CXX14
-
-UTL_NAMESPACE_END
-
-#  endif // ifdef UTL_BUILTIN_is_volatile
-
 #  define UTL_TRAIT_SUPPORTED_is_const 1
-#  define UTL_TRAIT_SUPPORTED_is_volatile 1
 
 #endif // ifdef UTL_USE_STD_TYPE_TRAITS
