+++ src/utl/public/utl/type_traits/utl_std_traits.h
@@ -4,6 +4,11 @@
 
 #include "utl/preprocessor/utl_namespace.h"
 #include "utl/preprocessor/utl_standard.h"
+#include "utl/type_traits/utl_add_const.h"
+#include "utl/type_traits/utl_add_cv.h"
+#include "utl/type_traits/utl_add_lvalue_reference.h"
+#include "utl/type_traits/utl_add_rvalue_reference.h"
+#include "utl/type_traits/utl_add_volatile.h"
 #include "utl/type_traits/utl_constants.h"
 #include "utl/type_traits/utl_decay.h"
 #include "utl/type_traits/utl_declval.h"
@@ -14,31 +19,40 @@
 #include "utl/type_traits/utl_is_class.h"
 #include "utl/type_traits/utl_is_complete.h"
 #include "utl/type_traits/utl_is_compound.h"
+#include "utl/type_traits/utl_is_const.h"
 #include "utl/type_traits/utl_is_enum.h"
 #include "utl/type_traits/utl_is_floating_point.h"
 #include "utl/type_traits/utl_is_function.h"
 #include "utl/type_traits/utl_is_fundamental.h"
 #include "utl/type_traits/utl_is_integral.h"
-#include "utl/type_traits/utl_is_member_x_pointer.h"
+#include "utl/type_traits/utl_is_lvalue_reference.h"
+#include "utl/type_traits/utl_is_member_function_pointer.h"
+#include "utl/type_traits/utl_is_member_object_pointer.h"
+#include "utl/type_traits/utl_is_member_pointer.h"
 #include "utl/type_traits/utl_is_null_pointer.h"
 #include "utl/type_traits/utl_is_object.h"
 #include "utl/type_traits/utl_is_pointer.h"
+#include "utl/type_traits/utl_is_reference.h"
+#include "utl/type_traits/utl_is_rvalue_reference.h"
 #include "utl/type_traits/utl_is_same.h"
 #include "utl/type_traits/utl_is_scalar.h"
+#include "utl/type_traits/utl_is_signed.h"
 #include "utl/type_traits/utl_is_union.h"
+#include "utl/type_traits/utl_is_unsigned.h"
 #include "utl/type_traits/utl_is_void.h"
+#include "utl/type_traits/utl_is_volatile.h"
 #include "utl/type_traits/utl_is_x_constructible.h"
 #include "utl/type_traits/utl_is_x_convertible.h"
-#include "utl/type_traits/utl_is_x_cv.h"
-#include "utl/type_traits/utl_is_x_reference.h"
-#include "utl/type_traits/utl_is_x_signed.h"
 #include "utl/type_traits/utl_logical_traits.h"
 #include "utl/type_traits/utl_modify_pointer.h"
-#include "utl/type_traits/utl_modify_x_cv.h"
-#include "utl/type_traits/utl_modify_x_reference.h"
 #include "utl/type_traits/utl_rank.h"
+#include "utl/type_traits/utl_remove_all_extents.h"
+#include "utl/type_traits/utl_remove_const.h"
+#include "utl/type_traits/utl_remove_cv.h"
 #include "utl/type_traits/utl_remove_cvref.h"
-#include "utl/type_traits/utl_remove_x_extent.h"
+#include "utl/type_traits/utl_remove_extent.h"
+#include "utl/type_traits/utl_remove_reference.h"
+#include "utl/type_traits/utl_remove_volatile.h"
 #include "utl/type_traits/utl_type_identity.h"
 #include "utl/type_traits/utl_void_t.h"
 
@@ -61,13 +75,12 @@ using ::std::has_virtual_destructor;
 
 template <typename... Ts>
 struct common_type : ::std::common_type<Ts...> {};
+template <typename... T>
+using common_type_t = typename common_type<T...>::type;
 
 using ::std::underlying_type;
 template <typename T>
 using underlying_type_t = typename underlying_type<T>::type;
-
-template <typename... T>
-using common_type_t = typename common_type<T...>::type;
 
 #ifdef UTL_BUILTIN_is_layout_compatible
 template <typename T, typename U>
