+++ src/utl/public/utl/utility/utl_forward.h
@@ -24,8 +24,8 @@ UTL_PRAGMA_WARN(
     "The current standard does not implement a constexpr forward, `UTL_USE_STD_forward` ignored")
 #  endif // defined(UTL_USE_STD_forward) && UTL_USE_STD_forward
 
-#  include "utl/type_traits/utl_is_x_reference.h"
-#  include "utl/type_traits/utl_modify_x_reference.h"
+#  include "utl/type_traits/utl_is_lvalue_reference.h"
+#  include "utl/type_traits/utl_remove_reference.h"
 
 UTL_NAMESPACE_BEGIN
 
@@ -58,8 +58,9 @@ UTL_NAMESPACE_END
 
 #else // defined(UTL_CXX23) && defined(UTL_USE_STD_forward_like) && UTL_USE_STD_forward_like
 
+#  include "utl/type_traits/utl_add_rvalue_reference.h"
 #  include "utl/type_traits/utl_copy_cvref.h"
-#  include "utl/type_traits/utl_modify_x_reference.h"
+#  include "utl/type_traits/utl_remove_reference.h"
 
 #  if defined(UTL_USE_STD_forward_like) && UTL_USE_STD_forward_like
 UTL_PRAGMA_WARN(
